fix notification

+++ ResellKH/ResellKH/src/main/java/com/example/resellkh/model/dto/AuthResponse.java
@@ -1,3 +1,4 @@
+// chouseangly/deployment/deployment-main/ResellKH/ResellKH/src/main/java/com/example/resellkh/model/dto/AuthResponse.java
 package com.example.resellkh.model.dto;
 
 import lombok.AllArgsConstructor;
@@ -8,7 +9,8 @@ import lombok.Data;
 public class AuthResponse {
     private String token;
     private String role;
-    private Integer userId;
+    // ✅ FIX: Changed from Integer to Long to match the Auth entity.
+    private Long userId;
     private String email;
     private String firstName;
     private String lastName;
