+++ ResellKH/ResellKH/src/main/java/com/example/resellkh/model/dto/FavouriteRequest.java
@@ -5,11 +5,12 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 public class FavouriteRequest {
-    private Integer userId;
-    private Integer productId;
+    private Long userId; // Changed from Integer
+    private Long productId; // Changed from Integer
     private LocalDateTime createdAt;
 }
