+++ ResellKH/ResellKH/src/main/java/com/example/resellkh/model/dto/RatingRequest.java
@@ -10,10 +10,9 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Data
 public class RatingRequest {
-    private int ratedUserId;
-    private int ratingUserId;
+    private Long ratedUserId; // Changed from int
+    private Long ratingUserId; // Changed from int
     private double score;
     private String comment;
     private LocalDateTime createdAt;
-
 }
