+++ ResellKH/ResellKH/src/main/java/com/example/resellkh/model/entity/Favourite.java
@@ -4,13 +4,14 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 public class Favourite {
     private Long favouriteId;
-    private Integer userId;
-    private Integer productId;
+    private Long userId; // Changed from Integer
+    private Long productId; // Changed from Integer
     private Product product;
     private LocalDateTime createdAt;
 }
