+++ ResellKH/ResellKH/src/main/java/com/example/resellkh/service/ProductHistoryService.java
@@ -3,8 +3,7 @@ import com.example.resellkh.model.entity.ProductHistory;
 import java.util.List;
 
 public interface ProductHistoryService {
-    void recordHistory(Integer productId, String message);
-    List<ProductHistory> getHistory(Integer productId);
-
+    void recordHistory(Long productId, String message);
+    List<ProductHistory> getHistory(Long productId);
     List<ProductHistory> getAllProductHistory();
 }
